lightspark/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
fastapi==0.116.1
h11==0.16.0
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT>=1.7.1,<2.10.0
PyMySQL==1.1.1
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
Pillow==11.3.0
python-jose[cryptography]==3.5.0
requests==2.32.5


lightspark==1.4.4
